apply static routes on change to master state
